requirements.txt
fastapi
uvicorn
httpie
requests
httpx
jinja2
python-multipart
pillow
bcrypt
python-dotenv
htmlmin
mercadopago
pyJWT